fix: playwright vsc extension not work
