--- tests/multiple-editor/first-editor.spec.ts
@@ -1,7 +1,0 @@
-import './setup-first.js';
-import '../basic.case.ts?first';
-import '../clipboard.case.ts?first';
-
-// console.log(test);
-import { setScope } from '../utils/playwright.js';
-setScope('');
